Fixed bug previous button not working
The slidesToShow variable is not even necessary here, so I removed that too.

I found this bug by creating a carousel with the following props:
- variableWidth: true
- infinite: false
- slidesToScroll: 3

I had 18 slides. When I clicked the next button for the last time, only the last slide showed, which is OK. But from then on the previous button stopped working. I looked up the source and found the bug. Fixed it and the problem went away.